requirements.txt
Flask
environs
pillow
pillow_heif
pypdf
